@@ -4,7 +4,7 @@
   "description": "A minimal Electron application with React and Express",
   "main": "main.js",
   "scripts": {
-    "dev": "concurrently -k \"npm run devWebpack\" \"npm run devApp\" \"npm run devServer\"",
+    "dev": "concurrently -k \"yarn run devWebpack\" \"yarn run devApp\" \"yarn run devServer\"",
     "devWebpack": "webpack-dev-server --content-base build/",
     "devApp": "electron main.js --dev",
     "devServer": "nodemon --watch backend backend/server.js",
@@ -25,7 +25,7 @@
     "babel-preset-react": "^6.24.1",
     "concurrently": "^3.5.0",
     "css-loader": "^0.28.4",
-    "electron": "~1.6.2",
+    "electron": "^1.7.5",
     "eslint": "^3.0.1",
     "eslint-plugin-react": "^5.2.2",
     "loglevel": "^1.4.1",
@@ -37,20 +37,27 @@
     "axios": "^0.16.2",
     "babel-polyfill": "^6.23.0",
     "body-parser": "^1.17.2",
+    "cookie-parser": "^1.4.3",
     "draft-js": "^0.10.1",
     "express": "^4.15.3",
+    "express-session": "^1.15.5",
     "fs": "^0.0.1-security",
+    "global": "^4.3.2",
     "immutable": "^3.8.1",
     "material-ui": "^0.18.7",
     "mongoose": "^4.11.4",
-    "passport": "^0.3.2",
+    "passport": "^0.4.0",
     "passport-facebook": "^2.1.1",
+    "passport-local": "^1.0.0",
+    "prop-types": "^15.5.10",
     "react": "^15.6.1",
     "react-color": "^2.13.4",
     "react-dom": "^15.6.1",
+    "react-redux": "^5.0.6",
     "react-router": "^4.1.2",
     "react-router-dom": "^4.1.2",
     "react-tap-event-plugin": "^2.0.1",
+    "redux": "^3.7.2",
     "socket.io": "^2.0.3",
     "style-loader": "^0.18.2"
   }
